Collapse h and hpp
